pull 05212019

+++ src/datasets.py
@@ -24,9 +24,10 @@ class IAM(data.Dataset):
     maybe_download('https://www.dropbox.com/sh/tdd0784neuv9ysh/AABm3gxtjQIZ2R9WZ-XR9Kpra?dl=0',
                    'iam_handwriting', root, 'folder')
     if exists(join(self.root,'words.tgz')):
-      os.makedirs(join(self.root, 'words'))
-      os.system('tar xvzf '+join(self.root, 'words.tgz')+' --directory '+join(self.root, 'words'))
-      os.system('rm '+join(self.root,'words.tgz'))
+      if not exists(join(self.root, 'words')):
+        os.makedirs(join(self.root, 'words'))
+        os.system('tar xvzf '+join(self.root, 'words.tgz')+' --directory '+join(self.root, 'words'))
+        os.system('rm '+join(self.root,'words.tgz'))
 
     # begin collecting all words in IAM dataset frm the words.txt summary file at the root of IAM directiory
     labelsFile = open(join(self.root,'words.txt'))
@@ -50,11 +51,15 @@ class IAM(data.Dataset):
       label = ' '.join(lineSplit[8:])
 
       # put sample into list
-      self.samples.append( (fileName, label) )
+      # qyk exclude empty images
+      if '---' not in label: # qyk: data clean
+        img_test=cv2.imread(fileName, cv2.IMREAD_GRAYSCALE) #qyk: data clean
+        if not (img_test is None or np.min(img_test.shape) <= 1): #qyk: data clean
+            self.samples.append( (fileName, label) ) #qyk
 
       # makes list of characters
-      chars = chars.union(set(list(label)))
-      self.charList = sorted(list(chars))
+            chars = chars.union(set(list(label)))
+    self.charList = sorted(list(chars))
 
   def __str__(self):
     return 'IAM words dataset. Data location: '+self.root+', Length: '+str(len(self))
@@ -68,6 +73,8 @@ class IAM(data.Dataset):
     # img = preprocess(cv2.imread(self.samples[i][0], cv2.IMREAD_GRAYSCALE),
     #                  args.imgsize, self.args, False, is_testing)
     img = cv2.imread(self.samples[idx][0], cv2.IMREAD_GRAYSCALE)
+    #print(self.samples[idx][0])#
+    #print(self.samples[idx][1]) #for debug purpose
     if self.transform:
       img = self.transform(img)
 
@@ -79,7 +86,8 @@ class EyDigitStrings(data.Dataset):
 
     self.transform = transform
     self.root = join(root, 'htr_assets/crowdsource/processed')
-
+    maybe_download(source_url='https://www.dropbox.com/s/dsg41kaajrfvfvj/htr_assets.zip?dl=0',
+                   filename='htr_assets', target_directory=root, filetype='zip') # qyk added, the source is yq's dropbox
     # custom dataset loader
     allfiles = glob(join(self.root, '**/*.jpg'), recursive=True)
     labels = [basename(f)[:-4] if basename(f).find('empty-')==-1 else '_' for f in allfiles] # if filename has 'empty-', then the ground truth is nothing
@@ -106,7 +114,7 @@ class EyDigitStrings(data.Dataset):
 
 class IRS(data.Dataset):
 
-  def __init__(self, root, transform=None):
+  def __init__(self, root='/root/datasets', transform=None):
 
     self.transform = transform
     self.root = join(root, 'irs_handwriting')
@@ -116,7 +124,11 @@ class IRS(data.Dataset):
     folder_depth = 2
     allfiles = glob(join(self.root, '**/'*folder_depth+'*.jpg'))
     labels = [basename(f)[:-4] for f in allfiles]
+    #print(labels[0])
     self.samples = list(zip(allfiles, labels))
+    # makes list of characters
+    chars = set.union(*[set(l) for l in labels])
+    self.charList = sorted(list(chars))
 
   def __len__(self):
     return len(self.samples)
@@ -133,7 +145,121 @@ class IRS(data.Dataset):
 
     return img, label
 
-dataroot = join(home,'datasets')
-iam = IAM(dataroot)
-ey = EyDigitStrings(dataroot)
-irs = IRS(dataroot)
+class PRT(data.Dataset):
+
+  def __init__(self, root='/root/datasets', transform=None):
+
+    self.transform = transform
+    self.root = join(root, 'img_print_100000_en')
+    maybe_download(source_url='https://www.dropbox.com/s/nbdu839yxfchdmv/img_print_100000_en.zip?dl=0',filename='img_print_100000_en', target_directory=root, filetype='zip') #'https://www.dropbox.com/s/cbhpy6clfi9a5lz/img_print_100000_clean.zip?dl=0'
+    #yq patch delete unrecognized non-english samples in linux
+    #os.system('find '+ root+' -maxdepth 1 -name "*.jpg" -type f -delete') find ./logs/examples -maxdepth 1 -name "*.log"
+    #if exists(join(root, 'img_print_100000_en')): os.system('mv ' + join(root, 'img_print_100000_en') + ' ' + self.root)
+
+    folder_depth = 0
+    allfiles = glob(join(self.root, '**/' * folder_depth + '*.jpg'))
+    #allfiles = [f for f in allfiles if len(basename(f))-4<=25 and len(basename(f))-4 >=1 and (not '#U' in f) and (not '---' in f)] # screen out non-recognized characters qyk
+    labels = [basename(f)[:-4] for f in allfiles]
+    self.samples = list(zip(allfiles, labels))
+    # makes list of characters
+    chars = set.union(*[set(l) for l in labels])
+    self.charList = sorted(list(chars))
+
+  def __len__(self):
+    return len(self.samples)
+
+  def __str__(self):
+    return 'Printing dataset. Data location: ' + self.root + ', Length: ' + str(len(self))
+
+  def __getitem__(self, idx):
+
+    label = self.samples[idx][1]
+    img = cv2.imread(self.samples[idx][0], cv2.IMREAD_GRAYSCALE)
+    if self.transform:
+      img = self.transform(img)
+
+    return img, label
+
+class PRT_WORD(data.Dataset):
+
+  def __init__(self, root='/root/datasets', transform=None):
+
+    self.transform = transform
+    self.root = join(root, 'img_print_single')
+    maybe_download(source_url='https://www.dropbox.com/s/xw8vd3n2jkz1n93/img_print_single.zip?dl=0',filename='img_print_single', target_directory=root, filetype='zip') #'https://www.dropbox.com/s/cbhpy6clfi9a5lz/img_print_100000_clean.zip?dl=0'
+    #yq patch delete unrecognized non-english samples in linux
+    #os.system('find '+ root+' -maxdepth 1 -name "*.jpg" -type f -delete') find ./logs/examples -maxdepth 1 -name "*.log"
+    #if exists(join(root, 'img_print_single')): os.system('mv ' + join(root, 'img_print_single') + ' ' + self.root)
+
+    folder_depth = 1
+    allfiles = glob(join(self.root, '**/' * folder_depth + '*.jpg'))
+    #allfiles = [f for f in allfiles if len(basename(f))-4<=25 and len(basename(f))-4 >=1 and (not '#U' in f) and (not '---' in f)] # screen out non-recognized characters qyk
+    labels = [basename(f)[:-4] for f in allfiles]
+    self.samples = list(zip(allfiles, labels))
+    # makes list of characters
+    chars = set.union(*[set(l) for l in labels])
+    self.charList = sorted(list(chars))
+
+  def __len__(self):
+    return len(self.samples)
+
+  def __str__(self):
+    return 'Printing dataset. Data location: ' + self.root + ', Length: ' + str(len(self))
+
+  def __getitem__(self, idx):
+
+    label = self.samples[idx][1]
+    img = cv2.imread(self.samples[idx][0], cv2.IMREAD_GRAYSCALE)
+    if self.transform:
+      img = self.transform(img)
+
+    return img, label
+def read_text(path):
+  with open(path,'r') as f:
+    txt=f.read().strip('\n')
+  return txt
+  
+
+class REAL(data.Dataset):
+
+  def __init__(self, root='/root/datasets', transform=None):
+
+    self.transform = transform
+    self.root = join(root, 'text_recognition')
+    maybe_download(source_url='https://www.dropbox.com/s/n1pq94xu9kpur1a/text_recognition.zip?dl=0',filename='text_recognition', target_directory=root, filetype='zip') #'https://www.dropbox.com/s/cbhpy6clfi9a5lz/img_print_100000_clean.zip?dl=0'
+    #yq patch delete unrecognized non-english samples in linux
+    #os.system('find '+ root+' -maxdepth 1 -name "*.jpg" -type f -delete') find ./logs/examples -maxdepth 1 -name "*.log"
+    #if exists(join(root, 'img_print_100000_en')): os.system('mv ' + join(root, 'img_print_100000_en') + ' ' + self.root)
+
+    #folder_depth = 0
+    allfiles = glob(join(self.root, 'imgs/' + '*.jpg'))
+    #allfiles = [f for f in allfiles if len(basename(f))-4<=25 and len(basename(f))-4 >=1 and (not '#U' in f) and (not '---' in f)] # screen out non-recognized characters qyk
+    labels = [read_text(f.replace('imgs','coord').replace('jpg','txt')) for f in allfiles]
+    print('real all: '+str(len(labels)))
+    all_samples = list(zip(allfiles, labels))
+    self.samples= [ sample for sample in all_samples if len(sample[1])<30] 
+    print('screened :'+str(len(self.samples)))
+    # makes list of characters
+    chars = set.union(*[set(l) for l in labels])
+    self.charList = sorted(list(chars))
+
+  def __len__(self):
+    return len(self.samples)
+
+  def __str__(self):
+    return 'Printing dataset. Data location: ' + self.root + ', Length: ' + str(len(self))
+
+  def __getitem__(self, idx):
+
+    label = self.samples[idx][1]
+    img = cv2.imread(self.samples[idx][0], cv2.IMREAD_GRAYSCALE)
+    if self.transform:
+      img = self.transform(img)
+
+    return img, label
+
+
+# dataroot = join(home,'datasets')
+# iam = IAM(dataroot)
+# ey = EyDigitStrings(dataroot)
+# irs = IRS(dataroot)
